reorganize code in persistence package
